requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.0
python-dotenv>=1.0.0
PyYAML>=6.0

# HTTP
aiohttp>=3.11.0

# Database
aiosqlite>=0.21.0

# AI / Discovery
chromadb>=1.5.0
playwright>=1.50.0
beautifulsoup4>=4.12.0

# Monitoring
prometheus-client>=0.22.0

# Security
# (built-in: unicodedata, re)

# Optional: Local LLM (LM Studio not required if bootstrap endpoints available)
# mlx>=0.22.0
# mlx-lm>=0.22.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Observability
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0
opentelemetry-instrumentation>=0.46b0
opentelemetry-instrumentation-fastapi>=0.46b0
opentelemetry-instrumentation-aiohttp-client>=0.46b0
opentelemetry-instrumentation-aiohttp-server>=0.46b0
opentelemetry-exporter-otlp>=1.25.0

# Utilities
cryptography>=42.0.0
PyJWT>=2.8.0
alembic>=1.13.0
